backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.13.0

# Database drivers
pymysql==1.1.0  # MySQL driver para SQLAlchemy
psycopg2-binary==2.9.9  # PostgreSQL driver (opcional)
# Si usas MySQL con mysqlclient (alternativa más rápida):
# mysqlclient==2.2.0

# Autenticación y seguridad (opcional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS
python-dotenv==1.0.0

# Dependency required by pydantic for email fields - version compatible with pydantic 2.x
email-validator==2.1.0

# Testing (opcional)
pytest==7.4.3
httpx==0.25.2
